After some thought, Nullification Pistol's effects EXCEPT scale work while a boss is alive again
Scale was the only effect that was truly broken at all with bosses using this thing. Removing the rest of the weapon's gimmick for bosses feels very awkward and unsatisfying. Personally I'm in the camp that that the scale really should be considered a feature too but that is beyond my jurisdiction at this point.